requirements.txt
asgiref==3.8.1
black==24.3.0
cfgv==3.4.0
click==8.1.7
distlib==0.3.9
Django==5.2.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
filelock==3.18.0
flake8==6.1.0
identify==2.6.12
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pre_commit==4.2.0
psycopg2-binary==2.9.6
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.9.0
python-dotenv==1.1.0
PyYAML==6.0.2
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.14.0
virtualenv==20.31.2
